Split assembly reference discovery
